requirements.txt
pytest
pytest-asyncio
fastapi
pynacl
